requirements.txt
appdata==2.1.2
astroid==2.14.1
black==23.1.0
click==8.0.3
colorama==0.4.6
colored==1.4.3
dill==0.3.6
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==2.6.2
pylint==2.16.1
SQLAlchemy==2.0.1
to-requirements.txt==1.3.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
Werkzeug==2.2.3
wrapt==1.14.1
WTForms==3.0.1
email_validator==1.3.1
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
